+++ src/context/SyntheticsStateContext/hooks/tradeHooks.ts
@@ -2,130 +2,13 @@ import { TradeMode, TradeType } from "domain/synthetics/trade";
 import { BigNumber } from "ethers";
 import { useMemo } from "react";
 import {
+  TokenTypeForSwapRoute,
   createTradeFlags,
-  makeSelectDecreasePositionAmounts,
-  makeSelectIncreasePositionAmounts,
-  makeSelectNextPositionValuesForDecrease,
   makeSelectNextPositionValuesForIncrease,
-  makeSelectSwapAmounts,
   makeSelectSwapRoutes,
   makeSelectTradeRatios,
 } from "../selectors/tradeSelectors";
 import { useSelector } from "../utils";
-
-export const useIncreasePositionAmounts = ({
-  collateralTokenAddress,
-  fixedAcceptablePriceImpactBps,
-  initialCollateralTokenAddress,
-  initialCollateralAmount,
-  leverage,
-  marketAddress,
-  positionKey,
-  strategy,
-  indexTokenAddress,
-  indexTokenAmount,
-  tradeMode,
-  tradeType,
-  triggerPrice,
-}: {
-  initialCollateralTokenAddress: string | undefined;
-  indexTokenAddress: string | undefined;
-  positionKey: string | undefined;
-  tradeMode: TradeMode;
-  tradeType: TradeType;
-  collateralTokenAddress: string | undefined;
-  marketAddress: string | undefined;
-  initialCollateralAmount: BigNumber;
-  indexTokenAmount: BigNumber;
-  leverage: BigNumber | undefined;
-  triggerPrice: BigNumber | undefined;
-  fixedAcceptablePriceImpactBps: BigNumber | undefined;
-  strategy: "leverageByCollateral" | "leverageBySize" | "independent";
-}) => {
-  const selector = useMemo(
-    () =>
-      makeSelectIncreasePositionAmounts({
-        collateralTokenAddress,
-        fixedAcceptablePriceImpactBps,
-        initialCollateralTokenAddress,
-        initialCollateralAmount,
-        leverage,
-        marketAddress,
-        positionKey,
-        strategy,
-        indexTokenAddress,
-        indexTokenAmount,
-        tradeMode,
-        tradeType,
-        triggerPrice,
-      }),
-    [
-      collateralTokenAddress,
-      fixedAcceptablePriceImpactBps,
-      indexTokenAddress,
-      indexTokenAmount,
-      initialCollateralAmount,
-      initialCollateralTokenAddress,
-      leverage,
-      marketAddress,
-      positionKey,
-      strategy,
-      tradeMode,
-      tradeType,
-      triggerPrice,
-    ]
-  );
-  return useSelector(selector);
-};
-
-export const useDecreasePositionAmounts = ({
-  collateralTokenAddress,
-  marketAddress,
-  positionKey,
-  tradeMode,
-  tradeType,
-  triggerPrice,
-  closeSizeUsd,
-  keepLeverage,
-  fixedAcceptablePriceImpactBps,
-}: {
-  positionKey: string | undefined;
-  tradeMode: TradeMode;
-  tradeType: TradeType;
-  collateralTokenAddress: string | undefined;
-  marketAddress: string | undefined;
-  triggerPrice: BigNumber | undefined;
-  closeSizeUsd: BigNumber | undefined;
-  fixedAcceptablePriceImpactBps: BigNumber | undefined;
-  keepLeverage: boolean;
-}) => {
-  const selector = useMemo(
-    () =>
-      makeSelectDecreasePositionAmounts({
-        collateralTokenAddress,
-        marketAddress,
-        positionKey,
-        tradeMode,
-        tradeType,
-        triggerPrice,
-        closeSizeUsd,
-        keepLeverage,
-        fixedAcceptablePriceImpactBps,
-      }),
-    [
-      closeSizeUsd,
-      collateralTokenAddress,
-      fixedAcceptablePriceImpactBps,
-      keepLeverage,
-      marketAddress,
-      positionKey,
-      tradeMode,
-      tradeType,
-      triggerPrice,
-    ]
-  );
-  return useSelector(selector);
-};
 
 export const useNextPositionValuesForIncrease = ({
   collateralTokenAddress,
@@ -141,6 +24,7 @@ export const useNextPositionValuesForIncrease = ({
   tradeMode,
   tradeType,
   triggerPrice,
+  tokenTypeForSwapRoute,
 }: {
   initialCollateralTokenAddress: string | undefined;
   indexTokenAddress: string | undefined;
@@ -155,6 +39,7 @@ export const useNextPositionValuesForIncrease = ({
   triggerPrice: BigNumber | undefined;
   fixedAcceptablePriceImpactBps: BigNumber | undefined;
   strategy: "leverageByCollateral" | "leverageBySize" | "independent";
+  tokenTypeForSwapRoute: TokenTypeForSwapRoute;
 }) => {
   const selector = useMemo(
     () =>
@@ -172,6 +57,7 @@ export const useNextPositionValuesForIncrease = ({
         tradeMode,
         tradeType,
         triggerPrice,
+        tokenTypeForSwapRoute,
       }),
     [
       collateralTokenAddress,
@@ -184,55 +70,7 @@ export const useNextPositionValuesForIncrease = ({
       marketAddress,
       positionKey,
       strategy,
-      tradeMode,
-      tradeType,
-      triggerPrice,
-    ]
-  );
-  return useSelector(selector);
-};
-
-export const useNextPositionValuesForDecrease = ({
-  closeSizeUsd,
-  collateralTokenAddress,
-  fixedAcceptablePriceImpactBps,
-  keepLeverage,
-  marketAddress,
-  positionKey,
-  tradeMode,
-  tradeType,
-  triggerPrice,
-}: {
-  closeSizeUsd: BigNumber | undefined;
-  collateralTokenAddress: string | undefined;
-  fixedAcceptablePriceImpactBps: BigNumber | undefined;
-  keepLeverage: boolean;
-  marketAddress: string | undefined;
-  positionKey: string | undefined;
-  tradeMode: TradeMode;
-  tradeType: TradeType;
-  triggerPrice: BigNumber | undefined;
-}) => {
-  const selector = useMemo(
-    () =>
-      makeSelectNextPositionValuesForDecrease({
-        collateralTokenAddress,
-        fixedAcceptablePriceImpactBps,
-        marketAddress,
-        positionKey,
-        tradeMode,
-        tradeType,
-        triggerPrice,
-        closeSizeUsd,
-        keepLeverage,
-      }),
-    [
-      closeSizeUsd,
-      collateralTokenAddress,
-      fixedAcceptablePriceImpactBps,
-      keepLeverage,
-      marketAddress,
-      positionKey,
+      tokenTypeForSwapRoute,
       tradeMode,
       tradeType,
       triggerPrice,
@@ -246,52 +84,6 @@ export const useSwapRoutes = (fromTokenAddress: string | undefined, toTokenAddre
     () => makeSelectSwapRoutes(fromTokenAddress, toTokenAddress),
     [fromTokenAddress, toTokenAddress]
   );
-  return useSelector(selector);
-};
-
-export const useSwapAmounts = ({
-  amountBy,
-  fromTokenAddress,
-  fromTokenAmount,
-  isWrapOrUnwrap,
-  toTokenAddress,
-  toTokenAmount,
-  tradeMode,
-  triggerRatioValue,
-}: {
-  fromTokenAddress: string | undefined;
-  toTokenAddress: string | undefined;
-  tradeMode: TradeMode;
-  isWrapOrUnwrap: boolean;
-  amountBy: "from" | "to" | undefined;
-  fromTokenAmount: BigNumber;
-  toTokenAmount: BigNumber;
-  triggerRatioValue: BigNumber | undefined;
-}) => {
-  const selector = useMemo(
-    () =>
-      makeSelectSwapAmounts({
-        amountBy,
-        fromTokenAddress,
-        fromTokenAmount,
-        isWrapOrUnwrap,
-        toTokenAddress,
-        toTokenAmount,
-        tradeMode,
-        triggerRatioValue,
-      }),
-    [
-      amountBy,
-      fromTokenAddress,
-      fromTokenAmount,
-      isWrapOrUnwrap,
-      toTokenAddress,
-      toTokenAmount,
-      tradeMode,
-      triggerRatioValue,
-    ]
-  );
-
   return useSelector(selector);
 };
